requirements.txt
ipywidgets
pandas
numpy
plotly
matplotlib
jupyter-dash
werkzeug==2.0.0
pyyaml
beautifulsoup4
requests
lxml